move / amalgamate types js_only
